third_party/blink/web_tests/svg/dynamic-updates/SVGForeignObjectElement-svgdom-height-prop-expected.txt

SVG 1.1 dynamic update tests
Test passed

Tests dynamic updates of the 'height' property of the SVGForeignObjectElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS foreignObjectElement.height.baseVal.value is 2
PASS foreignObjectElement.height.baseVal.value is 150
PASS successfullyParsed is true

TEST COMPLETE
